sfm, splat, recon: clippy-clean under --all-targets
The clippy gate's baseline is zero and these crates had 34 findings between
them, most older than the new code: index loops over one array (now
iterators), `len() > 0`, a same-type cast, and two complex types (`sfm`'s
features and verified pairs get names; the geometry pass returns the
existing `geometry::AuxGrad` instead of a bare tuple). Two are allowed with
a reason: the SH constants stay digit-for-digit identical to
`splat_sh.wgsl` (f32 rounds the extra digits away) and the 4x4 Jacobi
keeps the index form of its rotation formulas. Also the one finding each in
`backend-wgpu` (a no-op `u64::from`) and `model` (a doc list parsed as
a list item).

+++ crates/backend-wgpu/src/lib.rs
@@ -804,7 +804,7 @@ impl DeviceShared {
                     "wgpu error: {e}\n  {}\n  device limits: max_buffer_size {:.0} MiB, max_storage_buffer_binding_size {:.0} MiB",
                     tally.describe(),
                     mib(max_buffer),
-                    mib(u64::from(max_binding)),
+                    mib(max_binding),
                 );
             }));
         }
